requirements.txt
black
beautifulsoup4
EbookLib
ftfy
isort
pydub
google-cloud-texttospeech